Refactor creator staking pallet structure
